tons of tiny little edits and updates
I just ran through this guide with a new machine, recording my screen
as I went, so others could have some handholding through this bit if
they felt they needed it. Since I was on a new machine, and was
following these instructions as carefully as possible, I noticed a few
bits that could use more clarification.

So, I provided that clarification. Hope it helps